내용이 없는데 공유 시 튕기는 문제 해결 && 공유를 save to Files를 할 때 접근 권한 문제 및 파일이 통째로 날라가던 오류 해결

+++ NostalgiAlbum/AlbumScreenViewController/ToolBarController/Share/ShareViewController.swift
@@ -15,6 +15,7 @@ class ShareViewController: UIViewController, UIDocumentPickerDelegate {
     var filePath: URL?
     var existedAlbum : Bool!
     var albumCoverName : String!
+    var checkFileProvider : Bool!
     
     // MARK: - View Life Cycle
     override func viewDidLoad() {
@@ -46,7 +47,7 @@ class ShareViewController: UIViewController, UIDocumentPickerDelegate {
         //.nost file URL에서 .nost앞 앨범 이름만 따옴
         if !existedAlbum {
             do {
-                try unzipAlbumDirectory(AlbumCoverName: albumCoverName, shareFilePath: filePath!, deleteShareFile: true)
+                try unzipAlbumDirectory(AlbumCoverName: albumCoverName, shareFilePath: filePath!, checkFileProvider: checkFileProvider)
             } catch let error {
                 NSErrorHandling_Alert(error: error, vc: self)
                 // MARK: - Document에 저장된 album Directory 삭제
@@ -128,12 +129,13 @@ class ShareViewController: UIViewController, UIDocumentPickerDelegate {
     }
     
     @IBAction func closeInButtonTapped(_ sender: Any) {
-        do{
-            try FileManager.default.removeItem(at: filePath!)
-        } catch let error {
-            print("NSError Occur :: \(error)")
+        if !checkFileProvider {
+            do{
+                try FileManager.default.removeItem(at: filePath!)
+            } catch let error {
+                print("NSError Occur :: \(error)")
+            }
         }
-        
         self.dismiss(animated: false)
     }
